RDRP-857: tidy and move relevant tests
