+++ src/staging/staging_main.py
@@ -283,17 +283,10 @@ def run_staging(
         full_responses,
         secondary_full_responses,
         manual_outliers,
-        #ultfoc_mapper,
-        #itl_mapper,
-        #cellno_df,
         postcode_mapper,
-        #pg_num_alpha,
-        #sic_pg_alpha_mapper,
-        #sic_pg_utf_mapper,
         backdata,
         pg_detailed_mapper,
         itl1_detailed_mapper,
-        #ref_list_817_mapper,
         civil_defence_detailed_mapper,
         sic_division_detailed_mapper,
         manual_trim_df,
