+++ src/staging/validation.py
@@ -222,46 +222,6 @@ def combine_schemas_validate_full_df(
 
 @time_logger_wrap
 @exception_wrap
-def validate_ultfoc_df(df: pd.DataFrame) -> pd.DataFrame:
-    """
-    Validates ultfoc df:
-    1. Checks if the DataFrame has exactly two columns.
-    2. Checks if the column headers are 'ruref' and 'ultfoc'.
-    3. Checks the validity of values in the 'ultfoc' column.
-    Args:
-        df (pd.DataFrame): The input DataFrame containing 'ruref'
-        and 'ultfoc' columns.
-    """
-    try:
-        # Check DataFrame shape
-        if df.shape[1] != 2:
-            raise ValueError("Dataframe file must have exactly two columns")
-
-        # Check column headers
-        if list(df.columns) != ["ruref", "ultfoc"]:
-            raise ValueError("Column headers should be 'ruref' and 'ultfoc'")
-
-        # Check 'ultfoc' values are either 2 characters or 'nan'
-        def check_ultfoc(value):
-            if pd.isna(value):
-                return True
-            else:
-                return isinstance(value, str) and (len(value) == 2)
-
-        df["contents_check"] = df.apply(lambda row: check_ultfoc(row["ultfoc"]), axis=1)
-
-        # check any unexpected contents
-        if not df["contents_check"].all():
-            raise ValueError("Unexpected format within 'ultfoc' column contents")
-
-        df.drop(columns=["contents_check"], inplace=True)
-
-    except ValueError as ve:
-        raise ValueError("Foreign ownership mapper validation failed: " + str(ve))
-
-
-@time_logger_wrap
-@exception_wrap
 def validate_many_to_one(*args) -> pd.DataFrame:
     """
     Validates a many-to-one mapper DataFrame.
